Implement `q_delete_dup`
The function is implement using a way similar to finite-state machine
that has 2 state as following.
1. Duplication detection
2. Duplicated nodes removal
The queue are iterated and detected if the value of current node equals to
the value of the next node. If duplicated nodes are detected, it goes to
state 2 and start removing those follows, and change back to state 1
when all the nodes with same values are removed.
To keep the removal process simple and fast, the value of the first node
of the duplicated ones are taken into `del_val` and be freed at the end
of state 2 so that no other string copying process is needed.