Implement `q_swap` and reverse related functions
The swap process is implemented by a continual reptition of 'take node
and put it to the next' process.
The simple reverse function picks all the node iteratively to the front
of the linked list, and the K-reverse function splits out each sub-list
with K nodes in it, reverse it and put it back.