Refactor new_elem used in insertion functions
The length of input strings are originally checked if it might cause
integer overflow.
It is, however, unnecessary check since type size_t is practically
defined as unsigned long integer that contains 8 bytes long in linux,
so it can represent numbers up to 2^64-1. If a string actually contains
2^64 - 1 characters, it will use 2^64 - 1 byte (about 10^7 TB) that is
not able to store into any memory existing, so this check is not
necessary.

Besides, checking whether malloc for string in the element is checked
with early return to make the if statement shorter.